+++ lib/widgets/search/search_widget.dart
@@ -1,4 +1,5 @@
 import 'package:flutter/material.dart';
+import 'package:flutter_gen/gen_l10n/app_localizations.dart';
 import 'package:lotti/themes/theme.dart';
 
 // adapted from https://github.com/JohannesMilke/filter_listview_example
@@ -6,14 +7,14 @@ class SearchWidget extends StatefulWidget with PreferredSizeWidget {
   const SearchWidget({
     required this.text,
     required this.onChanged,
-    required this.hintText,
+    this.hintText,
     super.key,
     this.margin = const EdgeInsets.all(20),
   });
 
   final String text;
   final ValueChanged<String> onChanged;
-  final String hintText;
+  final String? hintText;
   final EdgeInsets margin;
 
   @override
@@ -28,6 +29,8 @@ class _SearchWidgetState extends State<SearchWidget> {
 
   @override
   Widget build(BuildContext context) {
+    final localizations = AppLocalizations.of(context)!;
+
     final styleActive = searchFieldStyle();
     final styleHint = searchFieldHintStyle();
 
@@ -46,18 +49,21 @@ class _SearchWidgetState extends State<SearchWidget> {
         controller: controller,
         decoration: InputDecoration(
           icon: Icon(Icons.search, color: style.color),
-          suffixIcon: GestureDetector(
-            child: Icon(
-              Icons.close_rounded,
-              color: style.color,
+          suffixIcon: Visibility(
+            visible: controller.text.isNotEmpty,
+            child: GestureDetector(
+              child: Icon(
+                Icons.close_rounded,
+                color: style.color,
+              ),
+              onTap: () {
+                controller.clear();
+                widget.onChanged('');
+                FocusScope.of(context).requestFocus(FocusNode());
+              },
             ),
-            onTap: () {
-              controller.clear();
-              widget.onChanged('');
-              FocusScope.of(context).requestFocus(FocusNode());
-            },
           ),
-          hintText: widget.hintText,
+          hintText: widget.hintText ?? localizations.searchHint,
           hintStyle: style,
           border: InputBorder.none,
         ),