+++ lib/widgets/habits/habits_filter.dart
@@ -5,6 +5,7 @@ import 'package:lotti/blocs/habits/habits_state.dart';
 import 'package:lotti/classes/entity_definitions.dart';
 import 'package:lotti/database/database.dart';
 import 'package:lotti/get_it.dart';
+import 'package:lotti/themes/theme.dart';
 import 'package:lotti/utils/color.dart';
 import 'package:pie_chart/pie_chart.dart';
 import 'package:tinycolor2/tinycolor2.dart';
@@ -45,7 +46,10 @@ class HabitsFilter extends StatelessWidget {
             return IconButton(
               padding: const EdgeInsets.all(5),
               icon: dataMap.isEmpty
-                  ? const Icon(Icons.filter_alt)
+                  ? Icon(
+                      Icons.filter_alt_off_outlined,
+                      color: styleConfig().secondaryTextColor,
+                    )
                   : PieChart(
                       dataMap: dataMap,
                       animationDuration: const Duration(milliseconds: 800),